Fix draco compilation on gcc11
The previous version breaks on Fedora 34, and likely other distros
with recent compilers due to not including cstddef and limits headers.

This is already fixed upstream, but only in master. And looking at
google/draco#704 there may be compatibility issues with using the
latest code.

So for now we're upgrading from 1.3.3 to 1.3.6 and adding the fix.